backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
openai==1.3.7
sqlglot==27.6.0
redis==5.0.1
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
xlsxwriter==3.1.9
pandas>=2.1.4
numpy>=1.25.2
aiofiles==23.2.1
httpx==0.25.2
